refactor(routes): extract model/openrouter/claude-plans routes to agents-models.ts
Moves /api/models/available, /api/openrouter/manual, /api/openrouter/models,
and /api/claude-plans handlers into a dedicated tryHandleAgentsModels function.

tryHandleAgentsModels is called first in tryHandleAgents (sorrend preserved).

agents.ts: 1709 -> 1612 lines | agents-models.ts: 112 lines new
